Add a menu option to open embeddings
This currently loads the embeddings, switches all the models to the
new embeddings and revalidates the inputs so far.

One glitch is that the validity color is not yet updated. So while it
is not possible to accept invalid input, the colors do not update yet
until the user makes a change in the query.